pyproject.toml
[project]
name = "yt-rag-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-functions>=1.24.0",
    "fastapi>=0.124.0",
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "ipykernel>=7.1.0",
    "lancedb>=0.25.3",
    "langchain-text-splitters>=1.1.0",
    "pandas>=2.3.3",
    "pydantic-ai>=1.28.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.1",
    "tiktoken>=0.12.0",
    "uvicorn>=0.38.0",
]
